Enables user to find the closest recycle bins
This project won 2nd prize in Lehman College NSBE 2015 Spring Hackathon.
It is a web app that enable users to find the closest recycle bins based
on their address. This project is implemented using JavaScript,
AngularJS, Bootstrap and HTML. Google maps API is used to locate
latitude and longitude of user's address and Sorcota API contains the
database of all recycle bins in NYC. Distance between latitude and
longitude of user and each recycle bin is compared. Then 10 closest
recycle bins are listed and shown map.